+++ ksql-rest-app/src/main/java/io/confluent/ksql/api/server/Server.java
@@ -161,6 +161,9 @@ public class Server {
   }
 
   synchronized SocketAddress getProxyTarget() {
+    if (jettyPort == -1) {
+      throw new IllegalStateException("jetty port not set");
+    }
     return SocketAddress.inetSocketAddress(jettyPort, "127.0.0.1");
   }
